% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveTM.R
\name{saveTM}
\alias{saveTM}
\title{Write data in a topicminer format}
\usage{
saveTM(texts, metadata, path)
}
\arguments{
\item{texts}{A character vector with your textual data.}

\item{metadata}{A data.frame object with your metadata.}

\item{path}{A string with path to the output directory.}
}
\description{
Write data in a topicminer format.
}
\examples{
saveTM(my_df$text, my_df[,c("user", "date", "location")], "./forTM/")
}
